refactor: 상품 목록 화면에서 데이터가 추가로 로드될 때 전체 상품을 반환하도록 변경
