+++ app/src/main/java/woowacourse/shopping/view/product/catalog/ProductCatalogViewModel.kt
@@ -7,15 +7,17 @@ import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.data.product.ProductRepository
 import woowacourse.shopping.data.product.ProductRepositoryImpl
 import woowacourse.shopping.domain.Product
-import woowacourse.shopping.view.PagedResult
 
 class ProductCatalogViewModel(
     private val repository: ProductRepository,
 ) : ViewModel() {
-    private val _products = MutableLiveData<PagedResult<Product>>()
-    val products: LiveData<PagedResult<Product>> = _products
+    private val items = mutableListOf<Product>()
+    private val _products = MutableLiveData<List<Product>>(items)
+    val products: LiveData<List<Product>> = _products
 
-    private var currentPage = 0
+    private var offset = FIRST_OFFSET
+    var hasNext = false
+        private set
 
     init {
         loadProducts()
@@ -25,19 +27,22 @@ class ProductCatalogViewModel(
         val result =
             repository.getPagedProducts(
                 PRODUCT_SIZE_LIMIT,
-                currentPage * PRODUCT_SIZE_LIMIT,
+                offset,
             )
-        _products.value = PagedResult(result.items, result.hasNext)
-        currentPage++
+        hasNext = result.hasNext
+        _products.value = products.value.orEmpty() + result.items
+        offset += result.items.size
     }
 
     private fun loadProducts() {
-        val result = repository.getPagedProducts(PRODUCT_SIZE_LIMIT, currentPage)
-        _products.value = PagedResult(result.items, result.hasNext)
-        currentPage++
+        val result = repository.getPagedProducts(PRODUCT_SIZE_LIMIT, offset)
+        hasNext = result.hasNext
+        _products.value = products.value.orEmpty() + result.items
+        offset += result.items.size
     }
 
     companion object {
+        private const val FIRST_OFFSET = 0
         private const val PRODUCT_SIZE_LIMIT = 20
 
         fun provideFactory(repository: ProductRepository = ProductRepositoryImpl()): ViewModelProvider.Factory =
