+++ app/src/main/res/layout/item_product.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="handler"
-            type="woowacourse.shopping.view.product.catalog.ProductViewHolder.EventHandler" />
+            type="woowacourse.shopping.view.product.catalog.adapter.ProductViewHolder.EventHandler" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
